Add files via upload
TASK 2 CREATE A BUSINESS CARD

Associated with ProgrammersTech



The Front side of the business card has a purple background with a floral design in the top left corner. The floral design is in a light purple color and is very detailed. The text "Favorite Desire" is in a large, light purple font and is centered on the card. Below the text is the website address "www.favaritedesire.com" in a smaller light purple font. The business card also has a gold design in the bottom right corner. The design is made up of lines that appear to be brushstrokes. 

Whereas the Back side of the business card is a dark purple color with a wavy, abstract shape and gold glitter accents on the right side. The company name is "FAVOURITE DESIRE" in a light purple font with the title "FOUNDER & CEO" underneath. The contact information is listed below with phone number, address, and email. Social media icons for Instagram, Facebook, Twitter, and WhatsApp are listed on the right side. The background color is a deep purple.