Add files via upload
TASK 1 DEVELOP A MARKETING BROCHURE
Associated with ProgrammersTech

The Tool is use is Adobe Photoshop

This is a brochure for a new fragrance called "Decent Aura" by JF Fragrances. The brochure features a modern design with a black and orange color scheme. The headline is "Introducing Decent Aura by JF Fragrances," and the subheading is "Captivate your senses." The brochure includes a variety of testimonials from satisfied customers, describing the fragrance as "enchanting," "sophisticated," and "simply exquisite." The brochure also highlights the fragrance's unique blend of citrus and floral notes, as well as its luxurious and timeless appeal. Overall, the brochure effectively communicates the key benefits and features of "Decent Aura," and invites potential customers to experience the fragrance for themselves.